feat(ssa):ssa database cache correct

+++ common/yak/ssa/ssadb/code.go
@@ -3,6 +3,7 @@ package ssadb
 import (
 	"context"
 	"encoding/json"
+
 	"github.com/davecgh/go-spew/spew"
 	"github.com/yaklang/yaklang/common/utils/memedit"
 	"github.com/yaklang/yaklang/common/yak/yaklib/codec"
@@ -104,12 +105,12 @@ func emptyIrCode() *IrCode {
 	return &IrCode{}
 }
 
-func RequireIrCode(DB *gorm.DB, program string) (uint, *IrCode) {
+func RequireIrCode(DB *gorm.DB, program string) (int64, *IrCode) {
 	db := DB.Model(&IrCode{})
 	ircode := emptyIrCode()
 	ircode.ProgramName = program
 	db.Create(ircode)
-	return ircode.ID, ircode
+	return int64(ircode.ID), ircode
 }
 
 func GetIrCodeById(db *gorm.DB, id int64) *IrCode {
